requirements.txt
dataclasses==0.7
Django==2.2.7
django-templated-mail==1.1.1
djangorestframework==3.10.3
djongo==1.2.36
djoser==2.0.3
gunicorn==20.0.0
pkg-resources==0.0.0
pymongo==3.9.0
pytz==2019.3
sqlparse==0.2.4
whitenoise==4.1.4
